Fix how tags are parsed/written (#4)

+++ doctor_test.go
@@ -58,7 +58,7 @@ func TestDoctor(t *testing.T) {
 type Dr struct {
 }
 
-func (d Dr) MessageMutator(n byte) Mutator {
+func (d Dr) MessageMutator(n uint64) Mutator {
 	return nil
 }
